+++ controllers/courses.js
@@ -48,8 +48,10 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
  * @route POST /api/v1/bootcamps/:bootcampId/courses
  * @access Private
  */
-exports.createCourse = asyncHandler(async (req, res, next) => {
+exports.addcourse = asyncHandler(async (req, res, next) => {
   const bootcamp = await Bootcamp.findById(req.params.bootcampId);
+  req.body.user = req.user.id;
+
   if (!bootcamp) {
     return next(
       new ErrorResponse(
@@ -58,7 +60,16 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
       )
     );
   }
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to add course to bootcamp ${bootcamp._id}`,
+        401
+      )
+    );
+  }
   const course = await Course.create({ ...req.body, bootcamp: bootcamp._id });
+
   res.status(201).json({
     success: true,
     data: course,
@@ -71,15 +82,27 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
  * @access Private
  */
 exports.updateCourse = asyncHandler(async (req, res, next) => {
-  const course = await Course.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  let course = await Course.findById(req.params.id);
   if (!course) {
     return next(
       new ErrorResponse(`Course not found with id of ${req.params.id}`, 404)
     );
   }
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update course ${course._id}`,
+        401
+      )
+    );
+  }
+
+  course = await Course.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
+  });
+
   res.status(200).json({
     success: true,
     data: course,
@@ -96,6 +119,15 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
   if (!course) {
     return next(
       new ErrorResponse(`Course not found with id of ${req.params.id}`, 404)
+    );
+  }
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course ${course._id}`,
+        401
+      )
     );
   }
   await course.deleteOne();
